Removed old controller code. Moved server to controllers. Moved tests to root.
