--- controllers/permits/applications/special-uses/getSchema.json
@@ -1,56 +1,0 @@
-{
-	"controlNumber": {"default":"", "intake":"accinstCn"},
-	"region": {"default":0, "intake":"middleLayer/region"},
-	"forest": {"default":0, "intake":"middleLayer/forest"},
-	"district": {"default":0, "intake":"middleLayer/district"},
-	"authorizingOfficerName": {"default":"", "intake":"authOfficerName"},
-	"authorizingOfficerTitle": {"default":"", "intake":"authOfficerTitle"},
-	"adminOrg": {"default":"", "intake":"adminOrg"},
-	"applicantInfo": {
-		"contactControlNumber":{"default":"", "intake":"addresses/contCn"},
-		"firstName": {"default":"", "intake":"holders/firstName"},
-		"lastName": {"default":"", "intake":"holders/lastName"},
-		"dayPhone": {
-			"areaCode": {"default":0, "intake":"phones/areaCode"},
-			"number": {"default":0, "intake":"phones/phoneNumber"},
-			"extension": {"default":0, "intake":"phones/extension"},
-			"type": {"default":"", "intake":"phones/phoneNumberType"}
-		},
-		"eveningPhone": {
-			"areaCode": {"default":0, "intake":"phones/areaCode"},
-			"number": {"default":0, "intake":"phones/phoneNumber"},
-			"extension": {"default":0, "intake":"phones/extension"},
-			"type": {"default":"", "intake":"phones/phoneNumberType"}
-		},
-		"emailAddress": {"default":"", "intake":"addresses/email"},
-		"mailingAddress": {"default":"", "intake":"addresses/address1"},
-		"mailingAddress2": {"default":"", "intake":"addresses/address2"},
-		"mailingCity": {"default":"", "intake":"addresses/cityName"},
-		"mailingState": {"default":"", "intake":"addresses/stateCode"},
-		"mailingZIP": {"default":0, "intake":"addresses/postalCode"},
-		"organizationName": {"default":"", "intake":"addresses/contName"},
-		"website": {"default":"", "intake":"middleLayer/website"},
-		"orgType": {"default":"", "intake":"holders/orgType"}
-	},
-	"noncommercialFields": {
-		"activityDescription": {"default":"", "intake":"middleLayer/activity_description"},
-		"locationDescription": {"default":"", "intake":"middleLayer/location_description"},
-		"startDateTime": {"default":"", "intake":"middleLayer/start_datetime"},
-		"endDateTime": {"default":"", "intake":"middleLayer/end_datetime"},
-		"numberParticipants": {"default":0, "intake":"middleLayer/number_participants"}
-	},
-	"tempOutfitterFields": {
-		"individualIsCitizen": {"default":false, "intake":"middleLayer/individual_is_citizen"},
-		"smallBusiness": {"default":false, "intake":"middleLayer/small_business"},
-		"activityDescription": {"default":"", "intake":"middleLayer/activity_description"},
-		"advertisingURL": {"default":"", "intake":"middleLayer/advertising_url"},
-		"advertisingDescription":{"default":"", "intake":"middleLayer/advertising_description"},
-		"clientCharges":{"default":"", "intake":"middleLayer/client_charges"},
-		"experienceList":{"default":"", "intake":"middleLayer/experience_list"},
-		"guideDocumentation": {"default":"", "intake":"middleLayer/"},
-		"acknowledgementOfRiskForm": {"default":"", "intake":"middleLayer/acknowledgementOfRiskForm"},
-		"insuranceCertificate": {"default":"", "intake":"middleLayer/insuranceCertificate"},
-		"goodStandingEvidence": {"default":"", "intake":"middleLayer/goodStandingEvidence"},
-		"operatingPlan": {"default":"", "intake":"middleLayer/operatingPlan"}
-	}
-}
