--- controllers/permits/applications/special-uses/postSchema.json
@@ -1,63 +1,0 @@
-{
-  "region": {"default":""},
-  "forest": {"default":""},
-  "district": {"default":""},
-  "authorizingOfficerName": {"default":""},
-  "authorizingOfficerTitle": {"default":""},
-  "securityId": {"default":""},
-  "managingOrg": {"default":""},
-  "adminOrg": {"default":""},
-  "sudsId": {"default":""},
-  "effectiveDate": {"default":""},
-  "status": {"default":"Accepted"},
-  "applicantInfo": {
-    "firstName": {"default":""},
-    "lastName": {"default":""},
-    "dayPhone": {
-      "areaCode": {"default":0},
-      "number": {"default":0},
-      "extension": {"default":""},
-      "type": {"default":""}
-    },
-    "eveningPhone": {
-      "areaCode": {"default":0},
-      "number": {"default":0},
-      "extension": {"default":""},
-      "type": {"default":""}
-    },
-    "emailAddress": {"default":""},
-    "mailingAddress": {"default":""},
-    "mailingAddress2": {"default":""},
-    "mailingCity": {"default":""},
-    "mailingState": {"default":""},
-    "mailingZIP": {"default":""},
-    "organizationName": {"default":""},
-    "website": {"default":""},
-    "orgType": {"default":""},
-    "countryName": {"default":"United States of America"},
-    "contName": {"default":""},
-    "contactType": {"default":""},
-    "addressType": {"default":"Main"} 
-  },
-  "type": {"default":""},
-  "noncommercialFields": {
-    "activityDescription": {"default":""},
-    "locationDescription": {"default":""},
-    "startDateTime": {"default":""},
-    "endDateTime": {"default":""},
-    "numberParticipants": {"default":0},
-    "purpose": {"default":""},
-    "crExempt": {"default":"Y"},
-    "formName": {"default":"FS-2700-3b"}
-  },
-  "tempOutfitterFields": {
-    "individualIsCitizen": {"default":false},
-    "smallBusiness": {"default": false},
-    "activityDescription": {"default":""},
-    "advertisingURL": {"default":""},
-    "advertisingDescription": {"default":""},
-    "clientCharges": {"default":""},
-    "experienceList": {"default":""},
-    "formName": {"default":"FS-2700-3f"}
-  }
-}
