--- routes/permits/applications/index.js
@@ -1,27 +1,0 @@
-/*
-
-  ___ ___       ___               _ _       _   ___ ___ 
- | __/ __|  ___| _ \___ _ _ _ __ (_) |_    /_\ | _ \_ _|
- | _|\__ \ / -_)  _/ -_) '_| '  \| |  _|  / _ \|  _/| | 
- |_| |___/ \___|_| \___|_| |_|_|_|_|\__| /_/ \_\_| |___|
-
-*/
-
-//*******************************************************************
-
-'use strict';
-
-//*******************************************************************
-// required modules
-
-const express = require('express');
-const router = express.Router();
-
-const specialUses = require('./special-uses');
-
-//*******************************************************************
-// router
-
-router.use('/special-uses', specialUses);
-
-module.exports = router;
