--- routes/permits/applications/special-uses/index.js
@@ -1,29 +1,0 @@
-/*
-
-  ___ ___       ___               _ _       _   ___ ___ 
- | __/ __|  ___| _ \___ _ _ _ __ (_) |_    /_\ | _ \_ _|
- | _|\__ \ / -_)  _/ -_) '_| '  \| |  _|  / _ \|  _/| | 
- |_| |___/ \___|_| \___|_| |_|_|_|_|\__| /_/ \_\_| |___|
-
-*/
-
-//*******************************************************************
-
-'use strict';
-
-//*******************************************************************
-// required modules
-
-const express = require('express');
-const router = express.Router();
-
-const commercial = require('./commercial');
-const noncommercial = require('./noncommercial');
-
-//*******************************************************************
-// router
-
-router.use('/commercial', commercial);
-router.use('/noncommercial', noncommercial);
-
-module.exports = router;
